+++ pants.toml
@@ -66,7 +66,6 @@ repo_id = "7775F8D5-FC58-4DBC-9302-D00AE4A1505F"
 
 [cli.alias]
 all-changed = "--changed-since=HEAD --changed-dependees=transitive"
-green = "fmt lint check"
 pyupgrade = "--backend-packages=pants.backend.python.lint.pyupgrade fmt"
